Move options interface to its own file
