+++ src/index.ts
@@ -20,42 +20,8 @@ import { fence } from './rules/fence.js';
 import { codeBlock } from './rules/codeBlock.js';
 import { image, redditImage } from './rules/image.js';
 import { reflink } from './rules/reflink.js';
+import { Options } from './options.js';
 import SimpleMarkdown from 'simple-markdown';
-
-export interface MediaMetadataImage {
-	e: 'Image';
-	m: 'image/png' | 'image/jpg';
-	s: AlbumEntry;
-	t?: string;
-	id: string;
-}
-
-export interface MediaMetadataGif {
-	e: 'AnimatedImage';
-	m: 'image/gif';
-	ext?: string;
-	s: {
-		y: number;
-		gif: string;
-		mp4?: string;
-		x: number;
-	};
-	t?: string;
-	id: string;
-}
-
-export interface AlbumEntry {
-	y: number;
-	x: number;
-	u: string;
-}
-
-export interface Options {
-	addTargetBlank?: boolean;
-	media_metadata?: {
-		[media_id: string]: MediaMetadataImage | MediaMetadataGif | undefined;
-	};
-}
 
 const rules = Object.assign({}, SimpleMarkdown.defaultRules, {
 	spoiler,
