+++ src/rules/url.ts
@@ -1,6 +1,6 @@
 import SimpleMarkdown from 'simple-markdown';
 import { SimpleMarkdownRule } from './ruleType.js';
-import { MediaMetadataImage } from '../index.js';
+import { MediaMetadataImage } from '../options.js';
 
 // Modifies original url regex to include www.
 export const urlRegex = /^((?:https?:\/\/|www\.)[^\s<]+[^<.,:;"')\]\s])/;
